FAIT: Travail pratique 2.

+++ src/Controller/UsersController.php
@@ -143,12 +143,16 @@ class UsersController extends AppController
     {
         $this->request->allowMethod(['post', 'delete']);
         $user = $this->Users->get($id);
-        if ($this->Users->delete($user)) {
-            $this->Flash->success(__('The user has been deleted.'));
-        } else {
-            $this->Flash->error(__('The user could not be deleted. Please, try again.'));
-        }
 
+        if($user->role == 'admin') {
+            $this->Flash->error(__('An administrator account could not be deleted.'));
+        } else {
+            if ($this->Users->delete($user)) {
+                $this->Flash->success(__('The user has been deleted.'));
+            } else {
+                $this->Flash->error(__('The user could not be deleted. Please, try again.'));
+            }
+        }
         return $this->redirect(['action' => 'index']);
     }
